requirements.txt
fbs==0.8.6
PyInstaller==3.4
pyqt5
Pillow
pyserial==3.4
sliplib==0.4.0
pytest==5.4.1
mock==3.0.5
bitarray-hardbyte==1.1.0
simpleaudio==1.0.4
wave==0.0.2
numpy
requests
pre-commit
